Attempting the beginnings of a major refactoring of fields.
This is the first commit toward a full re-work where fields are added
deliberately, and we eliminate entirely the KnownFieldInfo hierarchy.
Additionally, I have added a few helpers to enable better creation of
vector-related fields, and removed the old fields that had been unrolled for
each vector field.

Enzo is the testbed here.

--HG--
branch : yt-3.0